API_Rest/requirements.txt
fastapi==0.109.0 
uvicorn[standard]==0.27.0 
python-multipart==0.0.6 
tensorflow==2.16.1 
numpy==1.26.4 
pillow==10.2.0 
